+++ hotel/models.py
@@ -106,37 +106,3 @@ class ExtraService(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class Feedback(models.Model):
-    RATING_CHOICES = (
-        (1, '1'),
-        (2, '2'),
-        (3, '3'),
-        (4, '4'),
-        (5, '5'),
-    )
-
-    author = models.ForeignKey(
-        Client,
-        on_delete=models.CASCADE
-    )
-    created_at = models.DateTimeField(
-        auto_now_add=True,
-        verbose_name='Время создания'
-    )
-    rating = models.IntegerField(
-        choices=RATING_CHOICES,
-        default='5',
-        verbose_name='Оценка'
-    )
-    text = models.TextField(
-        verbose_name='Текст отзыва'
-    )
-
-    class Meta:
-        verbose_name = 'Отзывы'
-        verbose_name_plural = 'Отзывы'
-
-    def __str__(self):
-        return "Отзыв от {}".format(self.author)
