+++ hotel/views.py
@@ -42,29 +42,5 @@ class RoomBooking(generic.UpdateView):
         return super().form_valid(form)
 
 
-class FeedbackNew(generic.CreateView):
-    model = Feedback
-    template_name = 'hotel/feedback_new.html'
-    fields = ('rating', 'text')
-    success_url = '/'
-
-    def form_valid(self, form):
-        feedback = form.save(commit=False)
-        user = self.request.user
-        try:
-            client = Client.objects.get(user=user)
-        except Client.DoesNotExist:
-            raise forms.ValidationError("You are not our client!")
-        feedback.author = client
-        feedback.save()
-        return super().form_valid(form)
-
-
-class FeedbackList(generic.ListView):
-    model = Feedback
-    template_name = "hotel/feedback_list.html"
-    context_object_name = 'feedbacks'
-
-
 def components(request):
     return render(request, 'general/components.html')
